Fix SW output buffer mismatch: change out matrix C from int8_t to int16_t to correctly accommodate the 16-bit accumulated sums from the 8-bit matrix multiplication

+++ SW/mmult.cpp
@@ -13,7 +13,7 @@
 // Software implementation of Matrix Multiplication
 // The inputs are of the size (DATA_SIZE x DATA_SIZE)
 void m_softwareGold(const std::vector<int8_t> &in1,
-                    const std::vector<int8_t> &in2, std::vector<int8_t> &out) {
+                    const std::vector<int8_t> &in2, std::vector<int16_t> &out) {
   // Perform Matrix multiply Out = In1 x In2
   for (int i = 0; i < DATA_SIZE; i++) {
     for (int j = 0; j < DATA_SIZE; j++) {
@@ -46,17 +46,17 @@ int main(int argc, char **argv) {
   // core)
   auto bo_a = xrt::bo(device, matrix_size * sizeof(int8_t), mmult.group_id(0));
   auto bo_b = xrt::bo(device, matrix_size * sizeof(int8_t), mmult.group_id(1));
-  auto bo_c = xrt::bo(device, matrix_size * sizeof(int8_t), mmult.group_id(2));
+  auto bo_c = xrt::bo(device, matrix_size * sizeof(int16_t), mmult.group_id(2));
 
   // Para usarlos en el SW map a un puntero del entorno C++ tradicional
   int8_t *bo_a_map = bo_a.map<int8_t *>();
   int8_t *bo_b_map = bo_b.map<int8_t *>();
-  int8_t *bo_c_map = bo_c.map<int8_t *>();
+  int16_t *bo_c_map = bo_c.map<int16_t *>();
 
   // Inicializo arreglos en software para luego comparar
   std::vector<int8_t> sw_in1(matrix_size);
   std::vector<int8_t> sw_in2(matrix_size);
-  std::vector<int8_t> sw_out(matrix_size);
+  std::vector<int16_t> sw_out(matrix_size);
 
   // Lleno la memoria de la FPGA localmente y configuro mis arreglos de
   // verificacion
